Update Issue #5 Merge changes into the trunk.

+++ trunk/lib/Biodiverse/GUI/ProgressDialog.pm
@@ -6,6 +6,8 @@ package Biodiverse::GUI::ProgressDialog;
 
 use strict;
 use warnings;
+use 5.010;
+
 use Glib qw (TRUE FALSE);
 use Gtk2;
 use Gtk2::GladeXML;
@@ -32,60 +34,70 @@ sub new {
     my $text     = shift || $NULL_STRING;
     my $progress = shift || 0;
 
-    #  debug
-    #use Devel::StackTrace;
-    #my $trace = Devel::StackTrace->new;
-    #print $trace->as_string; # like carp
-
-    
     my $gui = Biodiverse::GUI::GUIManager->instance;
 
-    # Load the widgets from Glade's XML
+    # Load the widgets from Glade's XML - need a better method of detecting if we are run from a GUI
     my $glade_file = $gui->getGladeFile;
     Biodiverse::GUI::ProgressDialog::NotInGUI->throw
         if ! $glade_file;
 
-    my $dlgxml = Gtk2::GladeXML->new($glade_file, 'wndProgress');
-    my $dlg = $dlgxml->get_widget('wndProgress');
-    $dlg->set_transient_for( $gui->getWidget('wndMain') );
+    my $window = Gtk2::Window->new;
+    my $vbox   = Gtk2::VBox->new (undef, 10);
+    my $bar    = Gtk2::ProgressBar->new;
+    my $label  = Gtk2::Label->new;
     
-    $dlg->set_position('GTK_WIN_POS_MOUSE');
-
-    # Show the dialog
-    #$dlg->set_modal(1);
-    $dlg->show_all();
-    #$dlg->present();  #  raise to top
-
     # Make object
-    my $self = { dlgxml => $dlgxml, dlg => $dlg };
+    my $self = {
+        dlg          => $window,
+        label_widget => $label,
+        progress_bar => $bar,
+    };
     bless $self, $class;
-    
-    #$dlg->signal_connect (
-    #    'stop_pulsing'   => \&pulsate_stop,
-    #    $self,
-    #);
 
+    $window->set_transient_for( $gui->getWidget('wndMain') );
+    $window->set_default_size (300, -1);
+    $window->signal_connect ('delete-event' => \&destroy_callback, $self);
+    
+    $window->set_title ('Please wait...');
+    $label->set_line_wrap (1);
+    $label->set_markup($text);
+
+    $window->add ($vbox);
+    $vbox->pack_end ($bar,   1, 0, 0);
+    $vbox->pack_end ($label, 1, 0, 0);
+    $window->show_all;
+
+    
     $self->{progress_update_interval} = $progress_update_interval;
 
-    my $widget = $self->{dlgxml}->get_widget('label');
-    if ($widget) {
-        $widget->set_markup($text);
-    }
     $self->update ($text, 0, 1);
-
-    #$self->{last_update_time} = [gettimeofday];
 
     return $self;
 }
 
 sub destroy {
     my $self = shift;
-    
+    #say "Destroying progress bar";
     $self->pulsate_stop;
-    
-    $self->{dlg}->destroy();
-    
+
+    #  sometimes we have already been destroyed when this is called
+    if ($self->{dlg}) {
+        $self->{dlg}->destroy();
+    }
+
+    foreach my $key (keys %$self) {
+        #say "$key $self->{$key}";
+        $self->{$key} = undef;
+    }
+
     return;
+}
+
+#  wrapper for the destroy method
+sub destroy_callback {
+    my ($widget, $event, $self) = @_;
+    #say 'Destroy callback';
+    return $self->destroy;
 }
 
 sub update {  
@@ -113,21 +125,21 @@ sub update {
     #$self->{dlg}->present;  #  raise to top
 
     # update dialog
-    my $widget = $self->{dlgxml}->get_widget('label');
-    if ($widget) {
-        $widget->set_markup("<b>$text</b>");
+    my $label_widget = $self->{label_widget};
+    if ($label_widget) {
+        $label_widget->set_markup("<b>$text</b>");
     }
 
-    my $bar = $self->{dlgxml}->get_widget('progressbar');
-    
+    my $bar = $self->{progress_bar};
+
     if (not defined $bar) {
         Biodiverse::GUI::ProgressDialog::Cancel->throw(
             message  => "Progress bar closed, operation cancelled",
         );
     }
-    
+
     $self->{pulse} = 0;
-    
+
     $bar->set_fraction($progress);
 
     while (Gtk2->events_pending) { Gtk2->main_iteration(); }
@@ -146,7 +158,7 @@ sub pulsate {
     if (not defined $text) {
         $text = $NULL_STRING;
     }
-    
+
     #$self->update ($text, $progress);  #  We are avoiding pulsation for the moment...
     #return;
 
@@ -157,13 +169,15 @@ sub pulsate {
     }
     else {
         # update dialog
-        my $widget = $self->{dlgxml}->get_widget('label');
-        $widget->set_markup("<b>$text</b>") if $widget;
-        
-        my $bar = $self->{dlgxml}->get_widget('progressbar');
+        my $label_widget = $self->{label_widget};
+        if ($label_widget) {
+            $label_widget->set_markup("<b>$text</b>");
+        }
+
+        my $bar = $self->{bar};
         $bar->set_pulse_step ($progress);
         #$bar->pulse;
-        
+
         if (not $self->{pulse}) {  #  only set this if we aren't already pulsing
             print "Starting pulse\n";
             $self->{pulse} = 1;
@@ -171,15 +185,14 @@ sub pulsate {
             #my $x = Glib::Timeout->add(100, sub {pulse_progress_bar ( $self )});
             #my $y = $x;
         }
-        
+
         # process Gtk events - does this do the right thing?
         while (Gtk2->events_pending) { Gtk2->main_iteration(); }
-        
+
         #  bad idea this one.  It pulses without end.
         #Gtk2->main;
     }
-    
-    
+
     return;
 }
 
@@ -211,8 +224,7 @@ sub pulse_progress_bar {
     
     #print "[PROGRESS BAR] Stop pulsing\n";
     
-    return FALSE;
-    
+    return FALSE;    
 }
 
 1;
